+++ Drugstore/Dictionary/Forms output/PackingForm.cs
@@ -20,9 +20,8 @@ namespace Drugstore.Dictionary
         private void tsbAdd_Click(object sender, EventArgs e)
         {
             PackForm PForm = new PackForm();
-            PForm.ShowDialog();
-            //if (PForm.ShowDialog() == DialogResult.OK)
-                this.упаковкиTableAdapter.Fill(this.drugstoreDataSet.Упаковки);
+            if (PForm.ShowDialog() == DialogResult.OK)
+                упаковкиTableAdapter.Fill(drugstoreDataSet.Упаковки);
         }
 
         private void tsbEdit_Click(object sender, EventArgs e)
@@ -31,39 +30,35 @@ namespace Drugstore.Dictionary
             {
                 PackForm PForm = new PackForm(Convert.ToInt32(dataGridView1.CurrentRow.Cells[0].Value));
                 if (PForm.ShowDialog() == DialogResult.OK)
-                    this.упаковкиTableAdapter.Fill(this.drugstoreDataSet.Упаковки);
+                    упаковкиTableAdapter.Fill(drugstoreDataSet.Упаковки);
             }
         }
 
         private void tsbDelete_Click(object sender, EventArgs e)
         {
             if (dataGridView1.CurrentRow != null)
-            {
-                Packing.deleteId(Convert.ToInt32(dataGridView1.CurrentRow.Cells[0].Value));
-                this.упаковкиTableAdapter.Fill(this.drugstoreDataSet.Упаковки);
-            }
+                if (MessageBox.Show("Видалити запис?", "Попередження", MessageBoxButtons.YesNo) == DialogResult.Yes)
+                {
+                    Packing.deleteId(Convert.ToInt32(dataGridView1.CurrentRow.Cells[0].Value));
+                    упаковкиTableAdapter.Fill(drugstoreDataSet.Упаковки);
+                }
         }
 
         private void dataGridView1_CellDoubleClick(object sender, DataGridViewCellEventArgs e)
         {
             PackForm PForm = new PackForm(Convert.ToInt32(dataGridView1.CurrentRow.Cells[0].Value));
-            PForm.ShowDialog();
+            if (PForm.ShowDialog() == DialogResult.OK)
+                упаковкиTableAdapter.Fill(drugstoreDataSet.Упаковки);
         }
 
         private void PackingForm_Load(object sender, EventArgs e)
         {
-            this.упаковкиTableAdapter.Fill(this.drugstoreDataSet.Упаковки);
+            упаковкиTableAdapter.Fill(drugstoreDataSet.Упаковки);
         }
 
-        private void dataGridView1_CellContentClick(object sender, DataGridViewCellEventArgs e)
+        private void tsbRefresh_Click(object sender, EventArgs e)
         {
-
-        }
-
-        private void dataGridView1_CellDoubleClick_1(object sender, DataGridViewCellEventArgs e)
-        {
-            PackForm PForm = new PackForm(Convert.ToInt32(dataGridView1.CurrentRow.Cells[0].Value));
-            PForm.ShowDialog();
+            упаковкиTableAdapter.Fill(drugstoreDataSet.Упаковки);
         }
     }
 }
